add c submodule and remove overlap
