--- Cooklang.tab.c
@@ -1,1956 +1,0 @@
-/* A Bison parser, made by GNU Bison 3.3.2.  */
-
-/* Bison implementation for Yacc-like parsers in C
-
-   Copyright (C) 1984, 1989-1990, 2000-2015, 2018-2019 Free Software Foundation,
-   Inc.
-
-   This program is free software: you can redistribute it and/or modify
-   it under the terms of the GNU General Public License as published by
-   the Free Software Foundation, either version 3 of the License, or
-   (at your option) any later version.
-
-   This program is distributed in the hope that it will be useful,
-   but WITHOUT ANY WARRANTY; without even the implied warranty of
-   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-   GNU General Public License for more details.
-
-   You should have received a copy of the GNU General Public License
-   along with this program.  If not, see <http://www.gnu.org/licenses/>.  */
-
-/* As a special exception, you may create a larger work that contains
-   part or all of the Bison parser skeleton and distribute that work
-   under terms of your choice, so long as that work isn't itself a
-   parser generator using the skeleton or a modified version thereof
-   as a parser skeleton.  Alternatively, if you modify or redistribute
-   the parser skeleton itself, you may (at your option) remove this
-   special exception, which will cause the skeleton and the resulting
-   Bison output files to be licensed under the GNU General Public
-   License without this special exception.
-
-   This special exception was added by the Free Software Foundation in
-   version 2.2 of Bison.  */
-
-/* C LALR(1) parser skeleton written by Richard Stallman, by
-   simplifying the original so-called "semantic" parser.  */
-
-/* All symbols defined below should begin with yy or YY, to avoid
-   infringing on user name space.  This should be done even for local
-   variables, as they might otherwise be expanded by user macros.
-   There are some unavoidable exceptions within include files to
-   define necessary library symbols; they are noted "INFRINGES ON
-   USER NAME SPACE" below.  */
-
-/* Undocumented macros, especially those whose name start with YY_,
-   are private implementation details.  Do not rely on them.  */
-
-/* Identify Bison output.  */
-#define YYBISON 1
-
-/* Bison version.  */
-#define YYBISON_VERSION "3.3.2"
-
-/* Skeleton name.  */
-#define YYSKELETON_NAME "yacc.c"
-
-/* Pure parsers.  */
-#define YYPURE 0
-
-/* Push parsers.  */
-#define YYPUSH 0
-
-/* Pull parsers.  */
-#define YYPULL 1
-
-
-
-
-/* First part of user prologue.  */
-#line 2 "Cooklang.y" /* yacc.c:337  */
-
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-
-#include "include/CooklangParser.h"
-
-#define YYDEBUG 1
-
-int yylex();
-
-
-int yyerror ( Recipe * recipe, const char * s);
-
-extern void yyrestart( FILE * input_file );
-
-
-
-
-
-
-
-#line 94 "Cooklang.tab.c" /* yacc.c:337  */
-# ifndef YY_NULLPTR
-#  if defined __cplusplus
-#   if 201103L <= __cplusplus
-#    define YY_NULLPTR nullptr
-#   else
-#    define YY_NULLPTR 0
-#   endif
-#  else
-#   define YY_NULLPTR ((void*)0)
-#  endif
-# endif
-
-/* Enabling verbose error messages.  */
-#ifdef YYERROR_VERBOSE
-# undef YYERROR_VERBOSE
-# define YYERROR_VERBOSE 1
-#else
-# define YYERROR_VERBOSE 1
-#endif
-
-/* In a future release of Bison, this section will be replaced
-   by #include "Cooklang.tab.h".  */
-#ifndef YY_YY_COOKLANG_TAB_H_INCLUDED
-# define YY_YY_COOKLANG_TAB_H_INCLUDED
-/* Debug traces.  */
-#ifndef YYDEBUG
-# define YYDEBUG 0
-#endif
-#if YYDEBUG
-extern int yydebug;
-#endif
-
-/* Token type.  */
-#ifndef YYTOKENTYPE
-# define YYTOKENTYPE
-  enum yytokentype
-  {
-    WORD = 258,
-    MULTIWORD = 259,
-    UNIT = 260,
-    NUMBER = 261,
-    LCURL = 262,
-    RCURL = 263,
-    PUNC_CHAR = 264,
-    NL = 265,
-    TILDE = 266,
-    HWORD = 267,
-    ATWORD = 268,
-    METADATA = 269,
-    COMMENT = 270,
-    WHTS = 271
-  };
-#endif
-
-/* Value type.  */
-#if ! defined YYSTYPE && ! defined YYSTYPE_IS_DECLARED
-
-union YYSTYPE
-{
-#line 34 "Cooklang.y" /* yacc.c:352  */
-
-  char * string;
-  char character;
-  double number;
-
-#line 160 "Cooklang.tab.c" /* yacc.c:352  */
-};
-
-typedef union YYSTYPE YYSTYPE;
-# define YYSTYPE_IS_TRIVIAL 1
-# define YYSTYPE_IS_DECLARED 1
-#endif
-
-
-extern YYSTYPE yylval;
-
-int yyparse (Recipe * recipe);
-
-#endif /* !YY_YY_COOKLANG_TAB_H_INCLUDED  */
-
-
-
-#ifdef short
-# undef short
-#endif
-
-#ifdef YYTYPE_UINT8
-typedef YYTYPE_UINT8 yytype_uint8;
-#else
-typedef unsigned char yytype_uint8;
-#endif
-
-#ifdef YYTYPE_INT8
-typedef YYTYPE_INT8 yytype_int8;
-#else
-typedef signed char yytype_int8;
-#endif
-
-#ifdef YYTYPE_UINT16
-typedef YYTYPE_UINT16 yytype_uint16;
-#else
-typedef unsigned short yytype_uint16;
-#endif
-
-#ifdef YYTYPE_INT16
-typedef YYTYPE_INT16 yytype_int16;
-#else
-typedef short yytype_int16;
-#endif
-
-#ifndef YYSIZE_T
-# ifdef __SIZE_TYPE__
-#  define YYSIZE_T __SIZE_TYPE__
-# elif defined size_t
-#  define YYSIZE_T size_t
-# elif ! defined YYSIZE_T
-#  include <stddef.h> /* INFRINGES ON USER NAME SPACE */
-#  define YYSIZE_T size_t
-# else
-#  define YYSIZE_T unsigned
-# endif
-#endif
-
-#define YYSIZE_MAXIMUM ((YYSIZE_T) -1)
-
-#ifndef YY_
-# if defined YYENABLE_NLS && YYENABLE_NLS
-#  if ENABLE_NLS
-#   include <libintl.h> /* INFRINGES ON USER NAME SPACE */
-#   define YY_(Msgid) dgettext ("bison-runtime", Msgid)
-#  endif
-# endif
-# ifndef YY_
-#  define YY_(Msgid) Msgid
-# endif
-#endif
-
-#ifndef YY_ATTRIBUTE
-# if (defined __GNUC__                                               \
-      && (2 < __GNUC__ || (__GNUC__ == 2 && 96 <= __GNUC_MINOR__)))  \
-     || defined __SUNPRO_C && 0x5110 <= __SUNPRO_C
-#  define YY_ATTRIBUTE(Spec) __attribute__(Spec)
-# else
-#  define YY_ATTRIBUTE(Spec) /* empty */
-# endif
-#endif
-
-#ifndef YY_ATTRIBUTE_PURE
-# define YY_ATTRIBUTE_PURE   YY_ATTRIBUTE ((__pure__))
-#endif
-
-#ifndef YY_ATTRIBUTE_UNUSED
-# define YY_ATTRIBUTE_UNUSED YY_ATTRIBUTE ((__unused__))
-#endif
-
-/* Suppress unused-variable warnings by "using" E.  */
-#if ! defined lint || defined __GNUC__
-# define YYUSE(E) ((void) (E))
-#else
-# define YYUSE(E) /* empty */
-#endif
-
-#if defined __GNUC__ && ! defined __ICC && 407 <= __GNUC__ * 100 + __GNUC_MINOR__
-/* Suppress an incorrect diagnostic about yylval being uninitialized.  */
-# define YY_IGNORE_MAYBE_UNINITIALIZED_BEGIN \
-    _Pragma ("GCC diagnostic push") \
-    _Pragma ("GCC diagnostic ignored \"-Wuninitialized\"")\
-    _Pragma ("GCC diagnostic ignored \"-Wmaybe-uninitialized\"")
-# define YY_IGNORE_MAYBE_UNINITIALIZED_END \
-    _Pragma ("GCC diagnostic pop")
-#else
-# define YY_INITIAL_VALUE(Value) Value
-#endif
-#ifndef YY_IGNORE_MAYBE_UNINITIALIZED_BEGIN
-# define YY_IGNORE_MAYBE_UNINITIALIZED_BEGIN
-# define YY_IGNORE_MAYBE_UNINITIALIZED_END
-#endif
-#ifndef YY_INITIAL_VALUE
-# define YY_INITIAL_VALUE(Value) /* Nothing. */
-#endif
-
-
-#if ! defined yyoverflow || YYERROR_VERBOSE
-
-/* The parser invokes alloca or malloc; define the necessary symbols.  */
-
-# ifdef YYSTACK_USE_ALLOCA
-#  if YYSTACK_USE_ALLOCA
-#   ifdef __GNUC__
-#    define YYSTACK_ALLOC __builtin_alloca
-#   elif defined __BUILTIN_VA_ARG_INCR
-#    include <alloca.h> /* INFRINGES ON USER NAME SPACE */
-#   elif defined _AIX
-#    define YYSTACK_ALLOC __alloca
-#   elif defined _MSC_VER
-#    include <malloc.h> /* INFRINGES ON USER NAME SPACE */
-#    define alloca _alloca
-#   else
-#    define YYSTACK_ALLOC alloca
-#    if ! defined _ALLOCA_H && ! defined EXIT_SUCCESS
-#     include <stdlib.h> /* INFRINGES ON USER NAME SPACE */
-      /* Use EXIT_SUCCESS as a witness for stdlib.h.  */
-#     ifndef EXIT_SUCCESS
-#      define EXIT_SUCCESS 0
-#     endif
-#    endif
-#   endif
-#  endif
-# endif
-
-# ifdef YYSTACK_ALLOC
-   /* Pacify GCC's 'empty if-body' warning.  */
-#  define YYSTACK_FREE(Ptr) do { /* empty */; } while (0)
-#  ifndef YYSTACK_ALLOC_MAXIMUM
-    /* The OS might guarantee only one guard page at the bottom of the stack,
-       and a page size can be as small as 4096 bytes.  So we cannot safely
-       invoke alloca (N) if N exceeds 4096.  Use a slightly smaller number
-       to allow for a few compiler-allocated temporary stack slots.  */
-#   define YYSTACK_ALLOC_MAXIMUM 4032 /* reasonable circa 2006 */
-#  endif
-# else
-#  define YYSTACK_ALLOC YYMALLOC
-#  define YYSTACK_FREE YYFREE
-#  ifndef YYSTACK_ALLOC_MAXIMUM
-#   define YYSTACK_ALLOC_MAXIMUM YYSIZE_MAXIMUM
-#  endif
-#  if (defined __cplusplus && ! defined EXIT_SUCCESS \
-       && ! ((defined YYMALLOC || defined malloc) \
-             && (defined YYFREE || defined free)))
-#   include <stdlib.h> /* INFRINGES ON USER NAME SPACE */
-#   ifndef EXIT_SUCCESS
-#    define EXIT_SUCCESS 0
-#   endif
-#  endif
-#  ifndef YYMALLOC
-#   define YYMALLOC malloc
-#   if ! defined malloc && ! defined EXIT_SUCCESS
-void *malloc (YYSIZE_T); /* INFRINGES ON USER NAME SPACE */
-#   endif
-#  endif
-#  ifndef YYFREE
-#   define YYFREE free
-#   if ! defined free && ! defined EXIT_SUCCESS
-void free (void *); /* INFRINGES ON USER NAME SPACE */
-#   endif
-#  endif
-# endif
-#endif /* ! defined yyoverflow || YYERROR_VERBOSE */
-
-
-#if (! defined yyoverflow \
-     && (! defined __cplusplus \
-         || (defined YYSTYPE_IS_TRIVIAL && YYSTYPE_IS_TRIVIAL)))
-
-/* A type that is properly aligned for any stack member.  */
-union yyalloc
-{
-  yytype_int16 yyss_alloc;
-  YYSTYPE yyvs_alloc;
-};
-
-/* The size of the maximum gap between one aligned stack and the next.  */
-# define YYSTACK_GAP_MAXIMUM (sizeof (union yyalloc) - 1)
-
-/* The size of an array large to enough to hold all stacks, each with
-   N elements.  */
-# define YYSTACK_BYTES(N) \
-     ((N) * (sizeof (yytype_int16) + sizeof (YYSTYPE)) \
-      + YYSTACK_GAP_MAXIMUM)
-
-# define YYCOPY_NEEDED 1
-
-/* Relocate STACK from its old location to the new one.  The
-   local variables YYSIZE and YYSTACKSIZE give the old and new number of
-   elements in the stack, and YYPTR gives the new location of the
-   stack.  Advance YYPTR to a properly aligned location for the next
-   stack.  */
-# define YYSTACK_RELOCATE(Stack_alloc, Stack)                           \
-    do                                                                  \
-      {                                                                 \
-        YYSIZE_T yynewbytes;                                            \
-        YYCOPY (&yyptr->Stack_alloc, Stack, yysize);                    \
-        Stack = &yyptr->Stack_alloc;                                    \
-        yynewbytes = yystacksize * sizeof (*Stack) + YYSTACK_GAP_MAXIMUM; \
-        yyptr += yynewbytes / sizeof (*yyptr);                          \
-      }                                                                 \
-    while (0)
-
-#endif
-
-#if defined YYCOPY_NEEDED && YYCOPY_NEEDED
-/* Copy COUNT objects from SRC to DST.  The source and destination do
-   not overlap.  */
-# ifndef YYCOPY
-#  if defined __GNUC__ && 1 < __GNUC__
-#   define YYCOPY(Dst, Src, Count) \
-      __builtin_memcpy (Dst, Src, (Count) * sizeof (*(Src)))
-#  else
-#   define YYCOPY(Dst, Src, Count)              \
-      do                                        \
-        {                                       \
-          YYSIZE_T yyi;                         \
-          for (yyi = 0; yyi < (Count); yyi++)   \
-            (Dst)[yyi] = (Src)[yyi];            \
-        }                                       \
-      while (0)
-#  endif
-# endif
-#endif /* !YYCOPY_NEEDED */
-
-/* YYFINAL -- State number of the termination state.  */
-#define YYFINAL  2
-/* YYLAST -- Last index in YYTABLE.  */
-#define YYLAST   80
-
-/* YYNTOKENS -- Number of terminals.  */
-#define YYNTOKENS  17
-/* YYNNTS -- Number of nonterminals.  */
-#define YYNNTS  11
-/* YYNRULES -- Number of rules.  */
-#define YYNRULES  49
-/* YYNSTATES -- Number of states.  */
-#define YYNSTATES  71
-
-#define YYUNDEFTOK  2
-#define YYMAXUTOK   271
-
-/* YYTRANSLATE(TOKEN-NUM) -- Symbol number corresponding to TOKEN-NUM
-   as returned by yylex, with out-of-bounds checking.  */
-#define YYTRANSLATE(YYX)                                                \
-  ((unsigned) (YYX) <= YYMAXUTOK ? yytranslate[YYX] : YYUNDEFTOK)
-
-/* YYTRANSLATE[TOKEN-NUM] -- Symbol number corresponding to TOKEN-NUM
-   as returned by yylex.  */
-static const yytype_uint8 yytranslate[] =
-{
-       0,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     2,     2,     2,     2,
-       2,     2,     2,     2,     2,     2,     1,     2,     3,     4,
-       5,     6,     7,     8,     9,    10,    11,    12,    13,    14,
-      15,    16
-};
-
-#if YYDEBUG
-  /* YYRLINE[YYN] -- Source line where rule number YYN was defined.  */
-static const yytype_uint16 yyrline[] =
-{
-       0,    55,    55,    56,    61,    62,    70,    79,    83,    89,
-     102,   107,   108,   109,   110,   117,   124,   135,   136,   137,
-     141,   142,   148,   154,   160,   170,   174,   179,   185,   191,
-     195,   201,   205,   214,   219,   224,   229,   233,   239,   245,
-     252,   262,   276,   282,   289,   299,   312,   317,   322,   329
-};
-#endif
-
-#if YYDEBUG || YYERROR_VERBOSE || 1
-/* YYTNAME[SYMBOL-NUM] -- String name of the symbol SYMBOL-NUM.
-   First, the terminals, then, starting at YYNTOKENS, nonterminals.  */
-static const char *const yytname[] =
-{
-  "$end", "error", "$undefined", "WORD", "MULTIWORD", "UNIT", "NUMBER",
-  "LCURL", "RCURL", "PUNC_CHAR", "NL", "TILDE", "HWORD", "ATWORD",
-  "METADATA", "COMMENT", "WHTS", "$accept", "input", "line", "step",
-  "direction", "text_item", "amount", "cookware_amount", "cookware",
-  "ingredient", "timer", YY_NULLPTR
-};
-#endif
-
-# ifdef YYPRINT
-/* YYTOKNUM[NUM] -- (External) token number corresponding to the
-   (internal) symbol number NUM (which must be that of a token).  */
-static const yytype_uint16 yytoknum[] =
-{
-       0,   256,   257,   258,   259,   260,   261,   262,   263,   264,
-     265,   266,   267,   268,   269,   270,   271
-};
-# endif
-
-#define YYPACT_NINF -11
-
-#define yypact_value_is_default(Yystate) \
-  (!!((Yystate) == (-11)))
-
-#define YYTABLE_NINF -1
-
-#define yytable_value_is_error(Yytable_value) \
-  0
-
-  /* YYPACT[STATE-NUM] -- Index in YYTABLE of the portion describing
-     STATE-NUM.  */
-static const yytype_int8 yypact[] =
-{
-     -11,     2,   -11,   -11,   -11,   -11,   -11,   -11,     0,    58,
-      65,    -9,   -11,    17,   -11,    28,   -11,   -11,   -11,    10,
-      10,    32,   -11,    30,    30,    43,    49,   -11,    10,    10,
-      49,   -11,   -11,   -11,   -11,   -11,   -11,   -11,   -11,   -11,
-     -11,   -11,   -11,    52,    68,    70,   -11,    14,   -11,   -11,
-      31,    33,    35,   -11,    37,   -11,   -11,    42,   -11,    46,
-     -11,    48,   -11,   -11,   -11,   -11,   -11,   -11,   -11,   -11,
-     -11
-};
-
-  /* YYDEFACT[STATE-NUM] -- Default reduction number in state STATE-NUM.
-     Performed when YYTABLE does not specify something else to do.  Zero
-     means the default is an error.  */
-static const yytype_uint8 yydefact[] =
-{
-       2,     0,     1,    17,    18,    19,    20,     4,     0,    38,
-      42,     0,     3,     0,     7,    10,    12,    13,    11,    47,
-       0,     0,    46,    17,    18,     0,    14,    39,    17,    18,
-      15,    43,     6,     5,     9,     8,    21,    22,    23,    24,
-      16,    48,    49,     0,     0,     0,    25,     0,    40,    41,
-       0,     0,     0,    33,     0,    44,    45,     0,    29,     0,
-      31,     0,    27,    26,    36,    37,    35,    34,    30,    32,
-      28
-};
-
-  /* YYPGOTO[NTERM-NUM].  */
-static const yytype_int8 yypgoto[] =
-{
-     -11,   -11,   -11,   -11,    45,    15,   -10,    56,   -11,   -11,
-     -11
-};
-
-  /* YYDEFGOTO[NTERM-NUM].  */
-static const yytype_int8 yydefgoto[] =
-{
-      -1,     1,    12,    13,    14,    15,    22,    27,    16,    17,
-      18
-};
-
-  /* YYTABLE[YYPACT[STATE-NUM]] -- What to do in state STATE-NUM.  If
-     positive, shift that token.  If negative, reduce the rule whose
-     number is the opposite.  If YYTABLE_NINF, syntax error.  */
-static const yytype_uint8 yytable[] =
-{
-      31,    32,     2,    19,    20,     3,     4,    21,     5,    41,
-      42,     6,     7,     8,     9,    10,    11,    21,    55,    56,
-       3,     4,    63,     5,    26,    30,     6,    33,     8,     9,
-      10,    36,    37,    34,    38,    43,    44,    25,    45,    64,
-      46,    65,    39,    66,    40,    67,    50,    51,    47,    52,
-      68,    53,    36,    37,    69,    38,    70,    57,    35,    54,
-      58,    23,    24,    39,     5,    25,     0,     6,    28,    29,
-       0,     5,    21,    59,     6,    61,    60,     0,    62,    48,
-      49
-};
-
-static const yytype_int8 yycheck[] =
-{
-      10,    10,     0,     3,     4,     3,     4,     7,     6,    19,
-      20,     9,    10,    11,    12,    13,    14,     7,    28,    29,
-       3,     4,     8,     6,     9,    10,     9,    10,    11,    12,
-      13,     3,     4,    16,     6,     3,     4,     7,     6,     8,
-       8,     8,    14,     8,    16,     8,     3,     4,    16,     6,
-       8,     8,     3,     4,     8,     6,     8,     5,    13,    16,
-       8,     3,     4,    14,     6,     7,    -1,     9,     3,     4,
-      -1,     6,     7,     5,     9,     5,     8,    -1,     8,    23,
-      24
-};
-
-  /* YYSTOS[STATE-NUM] -- The (internal number of the) accessing
-     symbol of state STATE-NUM.  */
-static const yytype_uint8 yystos[] =
-{
-       0,    18,     0,     3,     4,     6,     9,    10,    11,    12,
-      13,    14,    19,    20,    21,    22,    25,    26,    27,     3,
-       4,     7,    23,     3,     4,     7,    22,    24,     3,     4,
-      22,    23,    10,    10,    16,    21,     3,     4,     6,    14,
-      16,    23,    23,     3,     4,     6,     8,    16,    24,    24,
-       3,     4,     6,     8,    16,    23,    23,     5,     8,     5,
-       8,     5,     8,     8,     8,     8,     8,     8,     8,     8,
-       8
-};
-
-  /* YYR1[YYN] -- Symbol number of symbol that rule YYN derives.  */
-static const yytype_uint8 yyr1[] =
-{
-       0,    17,    18,    18,    19,    19,    19,    20,    20,    20,
-      21,    21,    21,    21,    21,    21,    21,    22,    22,    22,
-      22,    22,    22,    22,    22,    23,    23,    23,    23,    23,
-      23,    23,    23,    24,    24,    24,    24,    24,    25,    25,
-      25,    25,    26,    26,    26,    26,    27,    27,    27,    27
-};
-
-  /* YYR2[YYN] -- Number of symbols on the right hand side of rule YYN.  */
-static const yytype_uint8 yyr2[] =
-{
-       0,     2,     0,     2,     1,     2,     2,     1,     2,     2,
-       1,     1,     1,     1,     2,     2,     2,     1,     1,     1,
-       1,     2,     2,     2,     2,     2,     3,     3,     4,     3,
-       4,     3,     4,     2,     3,     3,     3,     3,     1,     2,
-       3,     3,     1,     2,     3,     3,     2,     2,     3,     3
-};
-
-
-#define yyerrok         (yyerrstatus = 0)
-#define yyclearin       (yychar = YYEMPTY)
-#define YYEMPTY         (-2)
-#define YYEOF           0
-
-#define YYACCEPT        goto yyacceptlab
-#define YYABORT         goto yyabortlab
-#define YYERROR         goto yyerrorlab
-
-
-#define YYRECOVERING()  (!!yyerrstatus)
-
-#define YYBACKUP(Token, Value)                                    \
-  do                                                              \
-    if (yychar == YYEMPTY)                                        \
-      {                                                           \
-        yychar = (Token);                                         \
-        yylval = (Value);                                         \
-        YYPOPSTACK (yylen);                                       \
-        yystate = *yyssp;                                         \
-        goto yybackup;                                            \
-      }                                                           \
-    else                                                          \
-      {                                                           \
-        yyerror (recipe, YY_("syntax error: cannot back up")); \
-        YYERROR;                                                  \
-      }                                                           \
-  while (0)
-
-/* Error token number */
-#define YYTERROR        1
-#define YYERRCODE       256
-
-
-
-/* Enable debugging if requested.  */
-#if YYDEBUG
-
-# ifndef YYFPRINTF
-#  include <stdio.h> /* INFRINGES ON USER NAME SPACE */
-#  define YYFPRINTF fprintf
-# endif
-
-# define YYDPRINTF(Args)                        \
-do {                                            \
-  if (yydebug)                                  \
-    YYFPRINTF Args;                             \
-} while (0)
-
-/* This macro is provided for backward compatibility. */
-#ifndef YY_LOCATION_PRINT
-# define YY_LOCATION_PRINT(File, Loc) ((void) 0)
-#endif
-
-
-# define YY_SYMBOL_PRINT(Title, Type, Value, Location)                    \
-do {                                                                      \
-  if (yydebug)                                                            \
-    {                                                                     \
-      YYFPRINTF (stderr, "%s ", Title);                                   \
-      yy_symbol_print (stderr,                                            \
-                  Type, Value, recipe); \
-      YYFPRINTF (stderr, "\n");                                           \
-    }                                                                     \
-} while (0)
-
-
-/*-----------------------------------.
-| Print this symbol's value on YYO.  |
-`-----------------------------------*/
-
-static void
-yy_symbol_value_print (FILE *yyo, int yytype, YYSTYPE const * const yyvaluep, Recipe * recipe)
-{
-  FILE *yyoutput = yyo;
-  YYUSE (yyoutput);
-  YYUSE (recipe);
-  if (!yyvaluep)
-    return;
-# ifdef YYPRINT
-  if (yytype < YYNTOKENS)
-    YYPRINT (yyo, yytoknum[yytype], *yyvaluep);
-# endif
-  YYUSE (yytype);
-}
-
-
-/*---------------------------.
-| Print this symbol on YYO.  |
-`---------------------------*/
-
-static void
-yy_symbol_print (FILE *yyo, int yytype, YYSTYPE const * const yyvaluep, Recipe * recipe)
-{
-  YYFPRINTF (yyo, "%s %s (",
-             yytype < YYNTOKENS ? "token" : "nterm", yytname[yytype]);
-
-  yy_symbol_value_print (yyo, yytype, yyvaluep, recipe);
-  YYFPRINTF (yyo, ")");
-}
-
-/*------------------------------------------------------------------.
-| yy_stack_print -- Print the state stack from its BOTTOM up to its |
-| TOP (included).                                                   |
-`------------------------------------------------------------------*/
-
-static void
-yy_stack_print (yytype_int16 *yybottom, yytype_int16 *yytop)
-{
-  YYFPRINTF (stderr, "Stack now");
-  for (; yybottom <= yytop; yybottom++)
-    {
-      int yybot = *yybottom;
-      YYFPRINTF (stderr, " %d", yybot);
-    }
-  YYFPRINTF (stderr, "\n");
-}
-
-# define YY_STACK_PRINT(Bottom, Top)                            \
-do {                                                            \
-  if (yydebug)                                                  \
-    yy_stack_print ((Bottom), (Top));                           \
-} while (0)
-
-
-/*------------------------------------------------.
-| Report that the YYRULE is going to be reduced.  |
-`------------------------------------------------*/
-
-static void
-yy_reduce_print (yytype_int16 *yyssp, YYSTYPE *yyvsp, int yyrule, Recipe * recipe)
-{
-  unsigned long yylno = yyrline[yyrule];
-  int yynrhs = yyr2[yyrule];
-  int yyi;
-  YYFPRINTF (stderr, "Reducing stack by rule %d (line %lu):\n",
-             yyrule - 1, yylno);
-  /* The symbols being reduced.  */
-  for (yyi = 0; yyi < yynrhs; yyi++)
-    {
-      YYFPRINTF (stderr, "   $%d = ", yyi + 1);
-      yy_symbol_print (stderr,
-                       yystos[yyssp[yyi + 1 - yynrhs]],
-                       &yyvsp[(yyi + 1) - (yynrhs)]
-                                              , recipe);
-      YYFPRINTF (stderr, "\n");
-    }
-}
-
-# define YY_REDUCE_PRINT(Rule)          \
-do {                                    \
-  if (yydebug)                          \
-    yy_reduce_print (yyssp, yyvsp, Rule, recipe); \
-} while (0)
-
-/* Nonzero means print parse trace.  It is left uninitialized so that
-   multiple parsers can coexist.  */
-int yydebug;
-#else /* !YYDEBUG */
-# define YYDPRINTF(Args)
-# define YY_SYMBOL_PRINT(Title, Type, Value, Location)
-# define YY_STACK_PRINT(Bottom, Top)
-# define YY_REDUCE_PRINT(Rule)
-#endif /* !YYDEBUG */
-
-
-/* YYINITDEPTH -- initial size of the parser's stacks.  */
-#ifndef YYINITDEPTH
-# define YYINITDEPTH 200
-#endif
-
-/* YYMAXDEPTH -- maximum size the stacks can grow to (effective only
-   if the built-in stack extension method is used).
-
-   Do not make this value too large; the results are undefined if
-   YYSTACK_ALLOC_MAXIMUM < YYSTACK_BYTES (YYMAXDEPTH)
-   evaluated with infinite-precision integer arithmetic.  */
-
-#ifndef YYMAXDEPTH
-# define YYMAXDEPTH 10000
-#endif
-
-
-#if YYERROR_VERBOSE
-
-# ifndef yystrlen
-#  if defined __GLIBC__ && defined _STRING_H
-#   define yystrlen strlen
-#  else
-/* Return the length of YYSTR.  */
-static YYSIZE_T
-yystrlen (const char *yystr)
-{
-  YYSIZE_T yylen;
-  for (yylen = 0; yystr[yylen]; yylen++)
-    continue;
-  return yylen;
-}
-#  endif
-# endif
-
-# ifndef yystpcpy
-#  if defined __GLIBC__ && defined _STRING_H && defined _GNU_SOURCE
-#   define yystpcpy stpcpy
-#  else
-/* Copy YYSRC to YYDEST, returning the address of the terminating '\0' in
-   YYDEST.  */
-static char *
-yystpcpy (char *yydest, const char *yysrc)
-{
-  char *yyd = yydest;
-  const char *yys = yysrc;
-
-  while ((*yyd++ = *yys++) != '\0')
-    continue;
-
-  return yyd - 1;
-}
-#  endif
-# endif
-
-# ifndef yytnamerr
-/* Copy to YYRES the contents of YYSTR after stripping away unnecessary
-   quotes and backslashes, so that it's suitable for yyerror.  The
-   heuristic is that double-quoting is unnecessary unless the string
-   contains an apostrophe, a comma, or backslash (other than
-   backslash-backslash).  YYSTR is taken from yytname.  If YYRES is
-   null, do not copy; instead, return the length of what the result
-   would have been.  */
-static YYSIZE_T
-yytnamerr (char *yyres, const char *yystr)
-{
-  if (*yystr == '"')
-    {
-      YYSIZE_T yyn = 0;
-      char const *yyp = yystr;
-
-      for (;;)
-        switch (*++yyp)
-          {
-          case '\'':
-          case ',':
-            goto do_not_strip_quotes;
-
-          case '\\':
-            if (*++yyp != '\\')
-              goto do_not_strip_quotes;
-            else
-              goto append;
-
-          append:
-          default:
-            if (yyres)
-              yyres[yyn] = *yyp;
-            yyn++;
-            break;
-
-          case '"':
-            if (yyres)
-              yyres[yyn] = '\0';
-            return yyn;
-          }
-    do_not_strip_quotes: ;
-    }
-
-  if (! yyres)
-    return yystrlen (yystr);
-
-  return (YYSIZE_T) (yystpcpy (yyres, yystr) - yyres);
-}
-# endif
-
-/* Copy into *YYMSG, which is of size *YYMSG_ALLOC, an error message
-   about the unexpected token YYTOKEN for the state stack whose top is
-   YYSSP.
-
-   Return 0 if *YYMSG was successfully written.  Return 1 if *YYMSG is
-   not large enough to hold the message.  In that case, also set
-   *YYMSG_ALLOC to the required number of bytes.  Return 2 if the
-   required number of bytes is too large to store.  */
-static int
-yysyntax_error (YYSIZE_T *yymsg_alloc, char **yymsg,
-                yytype_int16 *yyssp, int yytoken)
-{
-  YYSIZE_T yysize0 = yytnamerr (YY_NULLPTR, yytname[yytoken]);
-  YYSIZE_T yysize = yysize0;
-  enum { YYERROR_VERBOSE_ARGS_MAXIMUM = 5 };
-  /* Internationalized format string. */
-  const char *yyformat = YY_NULLPTR;
-  /* Arguments of yyformat. */
-  char const *yyarg[YYERROR_VERBOSE_ARGS_MAXIMUM];
-  /* Number of reported tokens (one for the "unexpected", one per
-     "expected"). */
-  int yycount = 0;
-
-  /* There are many possibilities here to consider:
-     - If this state is a consistent state with a default action, then
-       the only way this function was invoked is if the default action
-       is an error action.  In that case, don't check for expected
-       tokens because there are none.
-     - The only way there can be no lookahead present (in yychar) is if
-       this state is a consistent state with a default action.  Thus,
-       detecting the absence of a lookahead is sufficient to determine
-       that there is no unexpected or expected token to report.  In that
-       case, just report a simple "syntax error".
-     - Don't assume there isn't a lookahead just because this state is a
-       consistent state with a default action.  There might have been a
-       previous inconsistent state, consistent state with a non-default
-       action, or user semantic action that manipulated yychar.
-     - Of course, the expected token list depends on states to have
-       correct lookahead information, and it depends on the parser not
-       to perform extra reductions after fetching a lookahead from the
-       scanner and before detecting a syntax error.  Thus, state merging
-       (from LALR or IELR) and default reductions corrupt the expected
-       token list.  However, the list is correct for canonical LR with
-       one exception: it will still contain any token that will not be
-       accepted due to an error action in a later state.
-  */
-  if (yytoken != YYEMPTY)
-    {
-      int yyn = yypact[*yyssp];
-      yyarg[yycount++] = yytname[yytoken];
-      if (!yypact_value_is_default (yyn))
-        {
-          /* Start YYX at -YYN if negative to avoid negative indexes in
-             YYCHECK.  In other words, skip the first -YYN actions for
-             this state because they are default actions.  */
-          int yyxbegin = yyn < 0 ? -yyn : 0;
-          /* Stay within bounds of both yycheck and yytname.  */
-          int yychecklim = YYLAST - yyn + 1;
-          int yyxend = yychecklim < YYNTOKENS ? yychecklim : YYNTOKENS;
-          int yyx;
-
-          for (yyx = yyxbegin; yyx < yyxend; ++yyx)
-            if (yycheck[yyx + yyn] == yyx && yyx != YYTERROR
-                && !yytable_value_is_error (yytable[yyx + yyn]))
-              {
-                if (yycount == YYERROR_VERBOSE_ARGS_MAXIMUM)
-                  {
-                    yycount = 1;
-                    yysize = yysize0;
-                    break;
-                  }
-                yyarg[yycount++] = yytname[yyx];
-                {
-                  YYSIZE_T yysize1 = yysize + yytnamerr (YY_NULLPTR, yytname[yyx]);
-                  if (yysize <= yysize1 && yysize1 <= YYSTACK_ALLOC_MAXIMUM)
-                    yysize = yysize1;
-                  else
-                    return 2;
-                }
-              }
-        }
-    }
-
-  switch (yycount)
-    {
-# define YYCASE_(N, S)                      \
-      case N:                               \
-        yyformat = S;                       \
-      break
-    default: /* Avoid compiler warnings. */
-      YYCASE_(0, YY_("syntax error"));
-      YYCASE_(1, YY_("syntax error, unexpected %s"));
-      YYCASE_(2, YY_("syntax error, unexpected %s, expecting %s"));
-      YYCASE_(3, YY_("syntax error, unexpected %s, expecting %s or %s"));
-      YYCASE_(4, YY_("syntax error, unexpected %s, expecting %s or %s or %s"));
-      YYCASE_(5, YY_("syntax error, unexpected %s, expecting %s or %s or %s or %s"));
-# undef YYCASE_
-    }
-
-  {
-    YYSIZE_T yysize1 = yysize + yystrlen (yyformat);
-    if (yysize <= yysize1 && yysize1 <= YYSTACK_ALLOC_MAXIMUM)
-      yysize = yysize1;
-    else
-      return 2;
-  }
-
-  if (*yymsg_alloc < yysize)
-    {
-      *yymsg_alloc = 2 * yysize;
-      if (! (yysize <= *yymsg_alloc
-             && *yymsg_alloc <= YYSTACK_ALLOC_MAXIMUM))
-        *yymsg_alloc = YYSTACK_ALLOC_MAXIMUM;
-      return 1;
-    }
-
-  /* Avoid sprintf, as that infringes on the user's name space.
-     Don't have undefined behavior even if the translation
-     produced a string with the wrong number of "%s"s.  */
-  {
-    char *yyp = *yymsg;
-    int yyi = 0;
-    while ((*yyp = *yyformat) != '\0')
-      if (*yyp == '%' && yyformat[1] == 's' && yyi < yycount)
-        {
-          yyp += yytnamerr (yyp, yyarg[yyi++]);
-          yyformat += 2;
-        }
-      else
-        {
-          yyp++;
-          yyformat++;
-        }
-  }
-  return 0;
-}
-#endif /* YYERROR_VERBOSE */
-
-/*-----------------------------------------------.
-| Release the memory associated to this symbol.  |
-`-----------------------------------------------*/
-
-static void
-yydestruct (const char *yymsg, int yytype, YYSTYPE *yyvaluep, Recipe * recipe)
-{
-  YYUSE (yyvaluep);
-  YYUSE (recipe);
-  if (!yymsg)
-    yymsg = "Deleting";
-  YY_SYMBOL_PRINT (yymsg, yytype, yyvaluep, yylocationp);
-
-  YY_IGNORE_MAYBE_UNINITIALIZED_BEGIN
-  YYUSE (yytype);
-  YY_IGNORE_MAYBE_UNINITIALIZED_END
-}
-
-
-
-
-/* The lookahead symbol.  */
-int yychar;
-
-/* The semantic value of the lookahead symbol.  */
-YYSTYPE yylval;
-/* Number of syntax errors so far.  */
-int yynerrs;
-
-
-/*----------.
-| yyparse.  |
-`----------*/
-
-int
-yyparse (Recipe * recipe)
-{
-    int yystate;
-    /* Number of tokens to shift before error messages enabled.  */
-    int yyerrstatus;
-
-    /* The stacks and their tools:
-       'yyss': related to states.
-       'yyvs': related to semantic values.
-
-       Refer to the stacks through separate pointers, to allow yyoverflow
-       to reallocate them elsewhere.  */
-
-    /* The state stack.  */
-    yytype_int16 yyssa[YYINITDEPTH];
-    yytype_int16 *yyss;
-    yytype_int16 *yyssp;
-
-    /* The semantic value stack.  */
-    YYSTYPE yyvsa[YYINITDEPTH];
-    YYSTYPE *yyvs;
-    YYSTYPE *yyvsp;
-
-    YYSIZE_T yystacksize;
-
-  int yyn;
-  int yyresult;
-  /* Lookahead token as an internal (translated) token number.  */
-  int yytoken = 0;
-  /* The variables used to return semantic value and location from the
-     action routines.  */
-  YYSTYPE yyval;
-
-#if YYERROR_VERBOSE
-  /* Buffer for error messages, and its allocated size.  */
-  char yymsgbuf[128];
-  char *yymsg = yymsgbuf;
-  YYSIZE_T yymsg_alloc = sizeof yymsgbuf;
-#endif
-
-#define YYPOPSTACK(N)   (yyvsp -= (N), yyssp -= (N))
-
-  /* The number of symbols on the RHS of the reduced rule.
-     Keep to zero when no symbol should be popped.  */
-  int yylen = 0;
-
-  yyssp = yyss = yyssa;
-  yyvsp = yyvs = yyvsa;
-  yystacksize = YYINITDEPTH;
-
-  YYDPRINTF ((stderr, "Starting parse\n"));
-
-  yystate = 0;
-  yyerrstatus = 0;
-  yynerrs = 0;
-  yychar = YYEMPTY; /* Cause a token to be read.  */
-  goto yysetstate;
-
-
-/*------------------------------------------------------------.
-| yynewstate -- push a new state, which is found in yystate.  |
-`------------------------------------------------------------*/
-yynewstate:
-  /* In all cases, when you get here, the value and location stacks
-     have just been pushed.  So pushing a state here evens the stacks.  */
-  yyssp++;
-
-
-/*--------------------------------------------------------------------.
-| yynewstate -- set current state (the top of the stack) to yystate.  |
-`--------------------------------------------------------------------*/
-yysetstate:
-  *yyssp = (yytype_int16) yystate;
-
-  if (yyss + yystacksize - 1 <= yyssp)
-#if !defined yyoverflow && !defined YYSTACK_RELOCATE
-    goto yyexhaustedlab;
-#else
-    {
-      /* Get the current used size of the three stacks, in elements.  */
-      YYSIZE_T yysize = (YYSIZE_T) (yyssp - yyss + 1);
-
-# if defined yyoverflow
-      {
-        /* Give user a chance to reallocate the stack.  Use copies of
-           these so that the &'s don't force the real ones into
-           memory.  */
-        YYSTYPE *yyvs1 = yyvs;
-        yytype_int16 *yyss1 = yyss;
-
-        /* Each stack pointer address is followed by the size of the
-           data in use in that stack, in bytes.  This used to be a
-           conditional around just the two extra args, but that might
-           be undefined if yyoverflow is a macro.  */
-        yyoverflow (YY_("memory exhausted"),
-                    &yyss1, yysize * sizeof (*yyssp),
-                    &yyvs1, yysize * sizeof (*yyvsp),
-                    &yystacksize);
-        yyss = yyss1;
-        yyvs = yyvs1;
-      }
-# else /* defined YYSTACK_RELOCATE */
-      /* Extend the stack our own way.  */
-      if (YYMAXDEPTH <= yystacksize)
-        goto yyexhaustedlab;
-      yystacksize *= 2;
-      if (YYMAXDEPTH < yystacksize)
-        yystacksize = YYMAXDEPTH;
-
-      {
-        yytype_int16 *yyss1 = yyss;
-        union yyalloc *yyptr =
-          (union yyalloc *) YYSTACK_ALLOC (YYSTACK_BYTES (yystacksize));
-        if (! yyptr)
-          goto yyexhaustedlab;
-        YYSTACK_RELOCATE (yyss_alloc, yyss);
-        YYSTACK_RELOCATE (yyvs_alloc, yyvs);
-# undef YYSTACK_RELOCATE
-        if (yyss1 != yyssa)
-          YYSTACK_FREE (yyss1);
-      }
-# endif
-
-      yyssp = yyss + yysize - 1;
-      yyvsp = yyvs + yysize - 1;
-
-      YYDPRINTF ((stderr, "Stack size increased to %lu\n",
-                  (unsigned long) yystacksize));
-
-      if (yyss + yystacksize - 1 <= yyssp)
-        YYABORT;
-    }
-#endif /* !defined yyoverflow && !defined YYSTACK_RELOCATE */
-
-  YYDPRINTF ((stderr, "Entering state %d\n", yystate));
-
-  if (yystate == YYFINAL)
-    YYACCEPT;
-
-  goto yybackup;
-
-
-/*-----------.
-| yybackup.  |
-`-----------*/
-yybackup:
-  /* Do appropriate processing given the current state.  Read a
-     lookahead token if we need one and don't already have one.  */
-
-  /* First try to decide what to do without reference to lookahead token.  */
-  yyn = yypact[yystate];
-  if (yypact_value_is_default (yyn))
-    goto yydefault;
-
-  /* Not known => get a lookahead token if don't already have one.  */
-
-  /* YYCHAR is either YYEMPTY or YYEOF or a valid lookahead symbol.  */
-  if (yychar == YYEMPTY)
-    {
-      YYDPRINTF ((stderr, "Reading a token: "));
-      yychar = yylex ();
-    }
-
-  if (yychar <= YYEOF)
-    {
-      yychar = yytoken = YYEOF;
-      YYDPRINTF ((stderr, "Now at end of input.\n"));
-    }
-  else
-    {
-      yytoken = YYTRANSLATE (yychar);
-      YY_SYMBOL_PRINT ("Next token is", yytoken, &yylval, &yylloc);
-    }
-
-  /* If the proper action on seeing token YYTOKEN is to reduce or to
-     detect an error, take that action.  */
-  yyn += yytoken;
-  if (yyn < 0 || YYLAST < yyn || yycheck[yyn] != yytoken)
-    goto yydefault;
-  yyn = yytable[yyn];
-  if (yyn <= 0)
-    {
-      if (yytable_value_is_error (yyn))
-        goto yyerrlab;
-      yyn = -yyn;
-      goto yyreduce;
-    }
-
-  /* Count tokens shifted since error; after three, turn off error
-     status.  */
-  if (yyerrstatus)
-    yyerrstatus--;
-
-  /* Shift the lookahead token.  */
-  YY_SYMBOL_PRINT ("Shifting", yytoken, &yylval, &yylloc);
-
-  /* Discard the shifted token.  */
-  yychar = YYEMPTY;
-
-  yystate = yyn;
-  YY_IGNORE_MAYBE_UNINITIALIZED_BEGIN
-  *++yyvsp = yylval;
-  YY_IGNORE_MAYBE_UNINITIALIZED_END
-
-  goto yynewstate;
-
-
-/*-----------------------------------------------------------.
-| yydefault -- do the default action for the current state.  |
-`-----------------------------------------------------------*/
-yydefault:
-  yyn = yydefact[yystate];
-  if (yyn == 0)
-    goto yyerrlab;
-  goto yyreduce;
-
-
-/*-----------------------------.
-| yyreduce -- do a reduction.  |
-`-----------------------------*/
-yyreduce:
-  /* yyn is the number of a rule to reduce with.  */
-  yylen = yyr2[yyn];
-
-  /* If YYLEN is nonzero, implement the default value of the action:
-     '$$ = $1'.
-
-     Otherwise, the following line sets YYVAL to garbage.
-     This behavior is undocumented and Bison
-     users should not rely upon it.  Assigning to YYVAL
-     unconditionally makes the parser a bit smaller, and it avoids a
-     GCC warning that YYVAL may be used uninitialized.  */
-  yyval = yyvsp[1-yylen];
-
-
-  YY_REDUCE_PRINT (yyn);
-  switch (yyn)
-    {
-        case 4:
-#line 61 "Cooklang.y" /* yacc.c:1652  */
-    {}
-#line 1299 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 5:
-#line 62 "Cooklang.y" /* yacc.c:1652  */
-    {
-      // after a step has been finished by a new line, have to add the step to the steplist
-      // and make a new step to accept directions
-      Step * newStep = createStep();
-
-      insertBack(recipe->stepList, newStep);
-      free((yyvsp[-1].string));
-    }
-#line 1312 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 6:
-#line 70 "Cooklang.y" /* yacc.c:1652  */
-    {
-      // add metadata to the recipe
-      addMetaData(recipe, (yyvsp[-1].string));
-      free((yyvsp[-1].string));
-    }
-#line 1322 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 7:
-#line 79 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = strdup((yyvsp[0].string));
-        free((yyvsp[0].string));
-      }
-#line 1331 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 8:
-#line 83 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = malloc(strlen((yyvsp[-1].string)) + strlen((yyvsp[0].string)) + 5);
-      sprintf((yyval.string), "%s %s", (yyvsp[-1].string), (yyvsp[0].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1342 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 9:
-#line 89 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = malloc(strlen((yyvsp[-1].string)) + strlen((yyvsp[0].string)) + 5);
-      sprintf((yyval.string), "%s %s", (yyvsp[-1].string), (yyvsp[0].string));
-
-      addDirection(recipe, "text", (yyvsp[0].string), NULL);
-
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1356 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 10:
-#line 102 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = strdup((yyvsp[0].string));
-      addDirection(recipe, "text", (yyvsp[0].string), NULL);
-      free((yyvsp[0].string));
-    }
-#line 1366 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 14:
-#line 110 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      addDirection(recipe, "cookware", (yyvsp[-1].string), NULL);
-      addDirection(recipe, "text", (yyvsp[0].string), NULL);
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1378 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 15:
-#line 117 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      addDirection(recipe, "ingredient", (yyvsp[-1].string), NULL);
-      addDirection(recipe, "text", (yyvsp[0].string), NULL);
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1390 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 16:
-#line 124 "Cooklang.y" /* yacc.c:1652  */
-    {
-    (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-    char * tempString = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-    addDirection(recipe, "text", tempString, NULL);
-    free(tempString);
-    free((yyvsp[-1].string));
-    free((yyvsp[0].string));
-  }
-#line 1403 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 19:
-#line 137 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = malloc(10);
-      sprintf((yyval.string), "%.3f", (yyvsp[0].number));
-    }
-#line 1412 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 21:
-#line 142 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1422 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 22:
-#line 148 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1432 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 23:
-#line 154 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = malloc(strlen((yyvsp[-1].string)) + 15);
-      sprintf((yyval.string), "%s %.3f", (yyvsp[-1].string), (yyvsp[0].number));
-      free((yyvsp[-1].string));
-    }
-#line 1442 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 24:
-#line 160 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1452 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 25:
-#line 170 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = malloc(5);
-      strcpy((yyval.string), "\0");
-    }
-#line 1461 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 26:
-#line 174 "Cooklang.y" /* yacc.c:1652  */
-    {
-    (yyval.string) = malloc(5);
-    strcpy((yyval.string), "\0");
-  }
-#line 1470 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 27:
-#line 179 "Cooklang.y" /* yacc.c:1652  */
-    {
-      // get string for amount
-      (yyval.string) = malloc(100);
-      sprintf((yyval.string), "%.3lf", (yyvsp[-1].number));
-    }
-#line 1480 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 28:
-#line 185 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = malloc(strlen((yyvsp[-1].string)) + 20);
-      sprintf((yyval.string), "%.3f %s", (yyvsp[-2].number), (yyvsp[-1].string));
-      free((yyvsp[-1].string));
-    }
-#line 1490 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 29:
-#line 191 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = (yyvsp[-1].string);
-    }
-#line 1498 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 30:
-#line 195 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-2].string), (yyvsp[-1].string));
-      free((yyvsp[-2].string));
-      free((yyvsp[-1].string));
-    }
-#line 1508 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 31:
-#line 201 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = (yyvsp[-1].string);
-    }
-#line 1516 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 32:
-#line 205 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-2].string), (yyvsp[-1].string));
-      free((yyvsp[-2].string));
-      free((yyvsp[-1].string));
-    }
-#line 1526 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 33:
-#line 214 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = malloc(5);
-        strcpy((yyval.string), "\0");
-      }
-#line 1535 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 34:
-#line 219 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = malloc(5);
-        strcpy((yyval.string), "\0");
-      }
-#line 1544 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 35:
-#line 224 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = malloc(100);
-        sprintf((yyval.string), "%.3f", (yyvsp[-1].number));
-      }
-#line 1553 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 36:
-#line 229 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = (yyvsp[-1].string);
-      }
-#line 1561 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 37:
-#line 233 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = (yyvsp[-1].string);
-      }
-#line 1569 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 38:
-#line 239 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = strdup((yyvsp[0].string));
-      addDirection(recipe, "cookware", (yyvsp[0].string), NULL);
-      free((yyvsp[0].string));
-    }
-#line 1579 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 39:
-#line 245 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      addDirection(recipe, "cookware", (yyvsp[-1].string), (yyvsp[0].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1590 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 40:
-#line 252 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addThreeStrings((yyvsp[-2].string), (yyvsp[-1].string), (yyvsp[0].string));
-      char * valueString = addTwoStrings((yyvsp[-2].string), (yyvsp[-1].string));
-      addDirection(recipe, "cookware", valueString, (yyvsp[0].string));
-      free(valueString);
-      free((yyvsp[-2].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1604 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 41:
-#line 262 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addThreeStrings((yyvsp[-2].string), (yyvsp[-1].string), (yyvsp[0].string));
-      char * valueString = addTwoStrings((yyvsp[-2].string), (yyvsp[-1].string));
-      addDirection(recipe, "cookware", valueString, (yyvsp[0].string));
-      free(valueString);
-      free((yyvsp[-2].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1618 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 42:
-#line 276 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = strdup((yyvsp[0].string));
-      addDirection(recipe, "ingredient", (yyvsp[0].string), NULL);
-      free((yyvsp[0].string));
-    }
-#line 1628 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 43:
-#line 282 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-      addDirection(recipe, "ingredient", (yyvsp[-1].string), (yyvsp[0].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1639 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 44:
-#line 289 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addThreeStrings((yyvsp[-2].string), (yyvsp[-1].string), (yyvsp[0].string));
-      char * valueString = addTwoStrings((yyvsp[-2].string), (yyvsp[-1].string));
-      addDirection(recipe, "ingredient", valueString, (yyvsp[0].string));
-      free(valueString);
-      free((yyvsp[-2].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1653 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 45:
-#line 299 "Cooklang.y" /* yacc.c:1652  */
-    {
-      (yyval.string) = addThreeStrings((yyvsp[-2].string), (yyvsp[-1].string), (yyvsp[0].string));
-      char * valueString = addTwoStrings((yyvsp[-2].string), (yyvsp[-1].string));
-      addDirection(recipe, "ingredient", valueString, (yyvsp[0].string));
-      free(valueString);
-      free((yyvsp[-2].string));
-      free((yyvsp[-1].string));
-      free((yyvsp[0].string));
-    }
-#line 1667 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 46:
-#line 312 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = strdup((yyvsp[0].string));
-        addDirection(recipe, "timer", NULL, (yyvsp[0].string));
-        free((yyvsp[0].string));
-      }
-#line 1677 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 47:
-#line 317 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = strdup((yyvsp[0].string));
-        addDirection(recipe, "timer", (yyvsp[0].string), NULL);
-        free((yyvsp[0].string));
-      }
-#line 1687 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 48:
-#line 322 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-        addDirection(recipe, "timer", (yyvsp[-1].string), (yyvsp[0].string));
-        free((yyvsp[-1].string));
-        free((yyvsp[0].string));
-      }
-#line 1698 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-  case 49:
-#line 329 "Cooklang.y" /* yacc.c:1652  */
-    {
-        (yyval.string) = addTwoStrings((yyvsp[-1].string), (yyvsp[0].string));
-        addDirection(recipe, "timer", (yyvsp[-1].string), (yyvsp[0].string));
-        free((yyvsp[-1].string));
-        free((yyvsp[0].string));
-      }
-#line 1709 "Cooklang.tab.c" /* yacc.c:1652  */
-    break;
-
-
-#line 1713 "Cooklang.tab.c" /* yacc.c:1652  */
-      default: break;
-    }
-  /* User semantic actions sometimes alter yychar, and that requires
-     that yytoken be updated with the new translation.  We take the
-     approach of translating immediately before every use of yytoken.
-     One alternative is translating here after every semantic action,
-     but that translation would be missed if the semantic action invokes
-     YYABORT, YYACCEPT, or YYERROR immediately after altering yychar or
-     if it invokes YYBACKUP.  In the case of YYABORT or YYACCEPT, an
-     incorrect destructor might then be invoked immediately.  In the
-     case of YYERROR or YYBACKUP, subsequent parser actions might lead
-     to an incorrect destructor call or verbose syntax error message
-     before the lookahead is translated.  */
-  YY_SYMBOL_PRINT ("-> $$ =", yyr1[yyn], &yyval, &yyloc);
-
-  YYPOPSTACK (yylen);
-  yylen = 0;
-  YY_STACK_PRINT (yyss, yyssp);
-
-  *++yyvsp = yyval;
-
-  /* Now 'shift' the result of the reduction.  Determine what state
-     that goes to, based on the state we popped back to and the rule
-     number reduced by.  */
-  {
-    const int yylhs = yyr1[yyn] - YYNTOKENS;
-    const int yyi = yypgoto[yylhs] + *yyssp;
-    yystate = (0 <= yyi && yyi <= YYLAST && yycheck[yyi] == *yyssp
-               ? yytable[yyi]
-               : yydefgoto[yylhs]);
-  }
-
-  goto yynewstate;
-
-
-/*--------------------------------------.
-| yyerrlab -- here on detecting error.  |
-`--------------------------------------*/
-yyerrlab:
-  /* Make sure we have latest lookahead translation.  See comments at
-     user semantic actions for why this is necessary.  */
-  yytoken = yychar == YYEMPTY ? YYEMPTY : YYTRANSLATE (yychar);
-
-  /* If not already recovering from an error, report this error.  */
-  if (!yyerrstatus)
-    {
-      ++yynerrs;
-#if ! YYERROR_VERBOSE
-      yyerror (recipe, YY_("syntax error"));
-#else
-# define YYSYNTAX_ERROR yysyntax_error (&yymsg_alloc, &yymsg, \
-                                        yyssp, yytoken)
-      {
-        char const *yymsgp = YY_("syntax error");
-        int yysyntax_error_status;
-        yysyntax_error_status = YYSYNTAX_ERROR;
-        if (yysyntax_error_status == 0)
-          yymsgp = yymsg;
-        else if (yysyntax_error_status == 1)
-          {
-            if (yymsg != yymsgbuf)
-              YYSTACK_FREE (yymsg);
-            yymsg = (char *) YYSTACK_ALLOC (yymsg_alloc);
-            if (!yymsg)
-              {
-                yymsg = yymsgbuf;
-                yymsg_alloc = sizeof yymsgbuf;
-                yysyntax_error_status = 2;
-              }
-            else
-              {
-                yysyntax_error_status = YYSYNTAX_ERROR;
-                yymsgp = yymsg;
-              }
-          }
-        yyerror (recipe, yymsgp);
-        if (yysyntax_error_status == 2)
-          goto yyexhaustedlab;
-      }
-# undef YYSYNTAX_ERROR
-#endif
-    }
-
-
-
-  if (yyerrstatus == 3)
-    {
-      /* If just tried and failed to reuse lookahead token after an
-         error, discard it.  */
-
-      if (yychar <= YYEOF)
-        {
-          /* Return failure if at end of input.  */
-          if (yychar == YYEOF)
-            YYABORT;
-        }
-      else
-        {
-          yydestruct ("Error: discarding",
-                      yytoken, &yylval, recipe);
-          yychar = YYEMPTY;
-        }
-    }
-
-  /* Else will try to reuse lookahead token after shifting the error
-     token.  */
-  goto yyerrlab1;
-
-
-/*---------------------------------------------------.
-| yyerrorlab -- error raised explicitly by YYERROR.  |
-`---------------------------------------------------*/
-yyerrorlab:
-  /* Pacify compilers when the user code never invokes YYERROR and the
-     label yyerrorlab therefore never appears in user code.  */
-  if (0)
-    YYERROR;
-
-  /* Do not reclaim the symbols of the rule whose action triggered
-     this YYERROR.  */
-  YYPOPSTACK (yylen);
-  yylen = 0;
-  YY_STACK_PRINT (yyss, yyssp);
-  yystate = *yyssp;
-  goto yyerrlab1;
-
-
-/*-------------------------------------------------------------.
-| yyerrlab1 -- common code for both syntax error and YYERROR.  |
-`-------------------------------------------------------------*/
-yyerrlab1:
-  yyerrstatus = 3;      /* Each real token shifted decrements this.  */
-
-  for (;;)
-    {
-      yyn = yypact[yystate];
-      if (!yypact_value_is_default (yyn))
-        {
-          yyn += YYTERROR;
-          if (0 <= yyn && yyn <= YYLAST && yycheck[yyn] == YYTERROR)
-            {
-              yyn = yytable[yyn];
-              if (0 < yyn)
-                break;
-            }
-        }
-
-      /* Pop the current state because it cannot handle the error token.  */
-      if (yyssp == yyss)
-        YYABORT;
-
-
-      yydestruct ("Error: popping",
-                  yystos[yystate], yyvsp, recipe);
-      YYPOPSTACK (1);
-      yystate = *yyssp;
-      YY_STACK_PRINT (yyss, yyssp);
-    }
-
-  YY_IGNORE_MAYBE_UNINITIALIZED_BEGIN
-  *++yyvsp = yylval;
-  YY_IGNORE_MAYBE_UNINITIALIZED_END
-
-
-  /* Shift the error token.  */
-  YY_SYMBOL_PRINT ("Shifting", yystos[yyn], yyvsp, yylsp);
-
-  yystate = yyn;
-  goto yynewstate;
-
-
-/*-------------------------------------.
-| yyacceptlab -- YYACCEPT comes here.  |
-`-------------------------------------*/
-yyacceptlab:
-  yyresult = 0;
-  goto yyreturn;
-
-
-/*-----------------------------------.
-| yyabortlab -- YYABORT comes here.  |
-`-----------------------------------*/
-yyabortlab:
-  yyresult = 1;
-  goto yyreturn;
-
-
-#if !defined yyoverflow || YYERROR_VERBOSE
-/*-------------------------------------------------.
-| yyexhaustedlab -- memory exhaustion comes here.  |
-`-------------------------------------------------*/
-yyexhaustedlab:
-  yyerror (recipe, YY_("memory exhausted"));
-  yyresult = 2;
-  /* Fall through.  */
-#endif
-
-
-/*-----------------------------------------------------.
-| yyreturn -- parsing is finished, return the result.  |
-`-----------------------------------------------------*/
-yyreturn:
-  if (yychar != YYEMPTY)
-    {
-      /* Make sure we have latest lookahead translation.  See comments at
-         user semantic actions for why this is necessary.  */
-      yytoken = YYTRANSLATE (yychar);
-      yydestruct ("Cleanup: discarding lookahead",
-                  yytoken, &yylval, recipe);
-    }
-  /* Do not reclaim the symbols of the rule whose action triggered
-     this YYABORT or YYACCEPT.  */
-  YYPOPSTACK (yylen);
-  YY_STACK_PRINT (yyss, yyssp);
-  while (yyssp != yyss)
-    {
-      yydestruct ("Cleanup: popping",
-                  yystos[*yyssp], yyvsp, recipe);
-      YYPOPSTACK (1);
-    }
-#ifndef yyoverflow
-  if (yyss != yyssa)
-    YYSTACK_FREE (yyss);
-#endif
-#if YYERROR_VERBOSE
-  if (yymsg != yymsgbuf)
-    YYSTACK_FREE (yymsg);
-#endif
-  return yyresult;
-}
-#line 338 "Cooklang.y" /* yacc.c:1918  */
-
-
-
-
-
-#include "lex.yy.c"
-
-
-
-int yyerror( Recipe * recipe, const char * s){
-  printf("\nError\n%s", s);
-  return 1;
-}
