+++ setup.py
@@ -7,12 +7,12 @@ setup(
         Extension(
             "cooklang",
             [
-                "src/CooklangExtension.c",
-                "src/CooklangParser.c",
-                "Cooklang.tab.c",
-                "src/LinkedListLib.c",
-                "src/CooklangRecipe.c",
-                "src/ShoppingListParser.c",
+                "CooklangExtension.c",
+                "cooklang-c/src/CooklangParser.c",
+                "cooklang-c/parserFiles/Cooklang.tab.c",
+                "cooklang-c/src/LinkedListLib.c",
+                "cooklang-c/src/CooklangRecipe.c",
+                "cooklang-c/src/ShoppingListParser.c",
             ],
         )
     ],
